requirements.txt
pybit==5.5.0
flask==3.0.3
pandas==2.2.1
flask-socketio==5.3.2
python-dotenv